refactor messagebox component to handle rendering messages
